Extract shared Pathname test helper
