+++ test/xml/test_document.rb
@@ -8,25 +8,6 @@ module Nokogiri
   module XML
     class TestDocument < Nokogiri::TestCase
       describe Nokogiri::XML::Document do
-        class TrackingPathname < Pathname
-          attr_reader :opened_io
-
-          def expand_path
-            self
-          end
-
-          def open(*args, &block)
-            if block
-              super(*args) do |io|
-                @opened_io = io
-                block.call(io)
-              end
-            else
-              @opened_io = super(*args)
-            end
-          end
-        end
-
         URI = if URI.const_defined?(:DEFAULT_PARSER)
           ::URI::DEFAULT_PARSER
         else
@@ -807,7 +788,7 @@ module Nokogiri
         def test_parse_can_take_pathnames
           assert_operator(File.size(XML_ATOM_FILE), :>, 4096) # file must be big enough to trip the read callback more than once
 
-          pathname = TrackingPathname.new(XML_ATOM_FILE)
+          pathname = Nokogiri::TestHelpers::TrackingPathname.new(XML_ATOM_FILE)
           doc = Nokogiri::XML.parse(pathname)
 
           # an arbitrary assertion on the structure of the document
